lesson2/requirements.txt
fastapi>=0.109.1,<0.110.0
uvicorn==0.23.2
openai==1.16.2
python-dotenv==1.0.1